feat: add document and stack architecture [incomplete]
The component architecture system was designed to enable developers to reuse code and also control the re-render flow of their messages, when to re-render and when not to re-render, but right now, we are struggling because we lack an evident architecture to hold these components.

Our current rendering architecture relies on lambdas to render the components, which means that to know differentiate between this old view and new view, we have to render both views and differentiate through what was changed and added, but that's not possible as rendering a view would re-render all components regardless of `shouldRerender` of the component.

Returning something like JSX.Element is also not viable as it leads to an issue similar to React where the developer experience is dropped. Currently, more investigations and research is needed to build a better architecture.